add updates to main.rb
