+++ main.rb
@@ -1,39 +1,36 @@
-require './src/app'
-
-def present_options
-  puts "
-  Choose an option:
-
-  1. List Items
-  2. Add a new Book
-  3. Add a new Music album
-  4. Add a new Movie
-  5. Add a new Game
-  6. Add a new Label
-  7. Exit"
-end
+# require './Menus/music_menu'
+# require './Menus/book_menu'
+require './src/Menus/game_menu'
 
 class Main
   def initialize
-    @app = App.new
+    @status = true
+    @welcome_message = [
+      "Welcome to the catalog app!\n",
+      'Please choose an option by entering a number:',
+      '1 - Book Menu',
+      '2 - Music Menu',
+      '3 - Games Menu',
+      '4 - Exit'
+    ]
   end
 
   def run
-    puts 'Welcome to the Catalog App!'
-    shutdown = false
-    until shutdown
-      present_options
-      option = gets.chomp.to_i
-      # handle option
-      if option.between?(1, 7)
-        shutdown = @app.main_processor(option)
-      else
-        puts 'Invalid option'
-      end
-      sleep(0.5)
-    end
+    while @status
+      @welcome_message.each { |i| puts i }
+      option = gets.chomp
 
-    puts 'Thank you for using the Catalogue!'
+      case option
+      # when '1' then BookMenu.new.run
+      # when '2' then MusicMenu.new.run
+      when '3' then GameMenu.new.run
+      when '4'
+        puts "Thank you for using the Catalogue!\n "
+        @status = false
+      else
+        puts "Sorry, you choose a wrong option\n "
+      end
+    end
   end
 end
