+++ src/MenuActions/create_game.rb
@@ -1,15 +1,17 @@
-require_relative '../Classes/game'
-require_relative '../Classes/game_author'
+require_relative '../classes/game'
+require_relative '../classes/game_author'
 
 class CreateGame
   def name_getter
     print('Enter the game name: ')
     gets.chomp
   end
+
   def author_getter
     print('Enter the author: ')
     gets.chomp
   end
+
   def multiplayer
     print('Is this game a Multiplayer(Y/N): ')
     player_method = gets.chomp.downcase
@@ -24,7 +26,8 @@ class CreateGame
   def publish_date
     print('Enter the publish date (YYYY-MM-DD): ')
     gets.chomp
-  end  
+  end
+
   def create_game(games)
     game = Game.new(publish_date, multiplayer, play_date, name_getter)
     author = Author.new(author_getter)
