+++ src/classes/game_author.rb
@@ -1,5 +1,5 @@
 require 'securerandom'
-require_relative 'item'
+require_relative '../item'
 
 # Create an Author class with an association to the Item class
 class Author
