+++ src/item.rb
@@ -20,6 +20,17 @@ class Item
 
   def add_genre(genre)
     @genre = genre
+    genre.items << self unless genre.items.include?(self)
+  end
+
+  def add_author(author)
+    @author = author
+    author.items << self unless author.items.include?(self)
+  end
+
+  def add_label(label)
+    @label = label
+    label.items << self unless label.items.include?(self)
   end
 
   attr_accessor :id, :genre, :author, :source, :label, :archived
